gap opening penalty includes gap extension penalty
In Flouri's tests, the gap opening penalty does not include the gap
extension penalty, and the optimal alignments contain two independent
gaps. Therefore, USEARCH and VSEARCH should return score values equal to
the scores indicated by Flouri, minus twice the gap extension
penalty (e.g., a score of -72 reported by Flouri corresponds to a score
of -70 with USEARCH and VSEARCH). The expected score values in the tests
are modified to take that into account.