Collecting CFLAGS to a central location
Collects the CFLAGS to a central location so that the debug flag `-g`,
the compiler optimization flag `-Ox`, and the warning flag `-Wx` would
be uniform accross all files.

`-Wsign-compare` is enabled by `-Wall`. Only its negation for cityhash
is kept. `-mcpu`, `-mtune`, `-msse2`, and `-msse3` are implied when
using `-march` and are removed.