initial test: write manpages in markdown, convert with pandoc
The goal is to address limitations of the current monolithic manpage:
- page is way too long,
- some parts are redundant,
- some parts are not developped enough,
- some commands don't have a list of available options,
- writen in groff (old-school format),
- current format not compatible with the web

The proposed solution is to write manpages in markdown and to convert
them with pandoc. A short manpage for vsearch (dispatch), and a
detailed manpage for each vsearch command.

Pros:
- easier format (pandoc-markdown),
- easy manpage generation,
- eliminates most text duplication (factorize redundant sections),
- usable online? (put generated markdown files on a website?)
- possibility to have man(7) documentation pages (e.g.; fastq format,
SHA1/MD5 hashing, quality values and offset, etc.)

Cons:
- adds new dependencies (perl and pandoc),
- pandoc-markdown, not real markdown,
- slightly less expressive than groff, but that should be ok

Preview:
cd ./man/commands/
bash visualize_manpage.sh vsearch-orient.1.md