clang-analyzer-unix.Stream
clang-tidy emits a warning: Stream pointer might be NULL

This should not happen as we first go through fatal() if the pointer
is NULL. However, future refactoring might create a read_sff_header()
function, and it would be good to clearly define the function
contract. Therefore, I decide to add a runtime DEBUG assertion to make
it clear that the function expects a non-NULL pointer.